+++ dfu/src/test/java/net/kyori/adventure/dfu/ShowItemCodecTest.java
@@ -21,19 +21,40 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.nbt.dfu;
+package net.kyori.adventure.dfu;
 
-import net.kyori.adventure.text.Component;
+import java.util.HashMap;
+import java.util.Map;
+import net.kyori.adventure.key.Key;
+import net.kyori.adventure.nbt.api.BinaryTagHolder;
+import net.kyori.adventure.text.event.DataComponentValue;
+import net.kyori.adventure.text.event.HoverEvent;
 import org.junit.jupiter.api.Test;
 
-public class SelectorComponentCodecTest extends AdventureCodecTest {
+public class ShowItemCodecTest extends AdventureCodecTest {
   @Test
   void test() {
-    assertComponentCodec(
-      Component.selector("selector")
+    assertCodec(
+      AdventureCodecs.SHOW_ITEM,
+      HoverEvent.ShowItem.showItem(
+        Key.key("stone"), 1
+      )
     );
-    assertComponentCodec(
-      Component.selector("selector", Component.text(","))
+    assertCodec(
+      AdventureCodecs.SHOW_ITEM,
+      HoverEvent.ShowItem.showItem(
+        Key.key("stone"), 1,
+        BinaryTagHolder.binaryTagHolder("{inValidPath: '123'}")
+      )
+    );
+    final Map<Key, DataComponentValue> componentValueMap = new HashMap<>();
+    componentValueMap.put(Key.key("in_valid_path"), BinaryTagHolder.binaryTagHolder("123"));
+    assertCodec(
+      AdventureCodecs.SHOW_ITEM,
+      HoverEvent.ShowItem.showItem(
+        Key.key("stone"), 1,
+        componentValueMap
+      )
     );
   }
 }
