+++ dfu/src/test/java/net/kyori/adventure/dfu/TranslatableComponentCodecTest.java
@@ -21,12 +21,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.adventure.nbt.dfu;
+package net.kyori.adventure.dfu;
 
 import com.google.gson.JsonElement;
 import com.mojang.serialization.JsonOps;
 import java.util.UUID;
-import net.kyori.adventure.dfu.AdventureCodecs;
 import net.kyori.adventure.key.Key;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TranslationArgument;
